外部設定ファイル分離完了まで
