+++ HookApp/Setting.settings
@@ -44,5 +44,8 @@
     <Setting Name="TITLE_UPDATE_NONE" Type="System.String" Scope="User">
       <Value Profile="(Default)">【最新バージョンです】</Value>
     </Setting>
+    <Setting Name="SYSTEM_SKIN_SETTING_FILE_NAME" Type="System.String" Scope="Application">
+      <Value Profile="(Default)">KeyDisplaySetting.yaml</Value>
+    </Setting>
   </Settings>
 </SettingsFile>
