+++ HookAppUnitTest/Models/AppSettingsModelTests.cs
@@ -21,13 +21,20 @@ namespace HookAppUnitTest.Models
             List<string> ansFolderList = new List<string> { "Default","Test2","Test3","yamlFormatInvalid","yamlReadTest"};
 
             ansFolderList.AssertIs(settingVm.SkinFolderNameList);
-
-            //Assert.IsTrue(settingVm.SkinFolderNameList.SequenceEqual(ansFolderList));
-            //Assert.IsTrue(settingVm.SelectedSkinFolderName == "Default");
         }
 
+        /// <summary>
+        /// Resourcesまでの絶対パスを取得できるかどうかテスト
+        /// </summary>
+        [TestMethod]
+        public void ObtainLocalPath()
+        {
+            AppSettingsModel settingVm = new AppSettingsModel();
 
+            var expected = @"C:\Users\Takahiro\source\repos\ktakahiro150397\GlobalHookWPF\HookAppUnitTest\bin\Debug\netcoreapp3.1\Resources\Default\KeyDisplaySetting.yaml";
 
+            settingVm.SelectedSkinSettingFilePath.AssertIs(expected);
 
+        }
     }
 }
